backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pymysql
python-dotenv
openai>=1.0.0
python-multipart
requests
cryptography